Work on low-level overrides for the global organization
Added back the override toggle on properties for the global organization, but not when a property is at the base level where that property is found since there is nothing else it can inherit from. For instance, the Digit Map property can't be toggled as overridden from the SoundPoint IP model page, but it can be toggled as overridden from the 335 model page since it has the manufacturer to inherit from.